requirements.txt
python-dotenv
requests
scapy